refactor: split weather widget into modular components
